# *: optimize query statements summary table by avoid decode plan for each row