<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Before this PR, query the  `statemens summary` table maybe very slow cause by decode plan, event the query doesn't need to read the `plan`.

### Bench

Generate 3000 statements to filling   `statemens summary` , and each SQL length is `10240 byte`.

1. Use [loadgen](https://github.com/crazycs520/load) to generated load:

   ```shell
   bin/loadgen payload  gen-stmt --stmt-cnt 3000 --query-len 10240 --host 172.16.5.40 --port 4019  --concurrency 10
   ```
   The generated SQL is look like this:
   
   ```sql
   select * from t_gen_stmt_1 where a = 1 or b = 1 or c = 1 
   union all
   select * from t_gen_stmt_2 where a = 2 or b = 2 or c = 2 
   union all
   ...
   ...
   ...
   ```
   
2. Wait generated statements(SQL) fill in the  `statemens summary`.

3. Bench the SQL: `select count(*) from information_schema.cluster_statements_summary_history;`

Bellow is the bench result compare with the master:

|                      | Master | This PR |
| ---------------------| ------ | ------- |
| Query duration       | 3.4s   | 7ms     |
| Query memory consume | 500MB  | < 1MB   |

Related TiDB memory consume metric is below:

Master:

![image](https://user-images.githubusercontent.com/26020263/124219309-af3e3400-db2e-11eb-946f-5f146b793273.png)

This PR:

![image](https://user-images.githubusercontent.com/26020263/124219635-2a074f00-db2f-11eb-972e-f8c2255eed51.png)




## # What is changed and how it works?

This PR add a special `stmtSummaryReader` to read the  `statemens summary` table data, and only `'decode'` the needed column, avoid decode plan for each table row.

Also, this PR implements `prune columns` for `LogicalMemTable` if the `mem table` is   `statemens summary` table. Otherwise, `stmtSummaryReader` still has to read all column since `LogicalMemTable` doesn't `prune column`.

### Related changes

- N/A

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
- Manual test

Side effects

- - N/A

### Release note <!-- bugfixes or new feature need a release note -->

- N/A